move sort_groups() to group_utils.py
